# spec/openapi.yaml
openapi: 3.1.0
info:
  title: SEAM API
  version: 0.1.0
  description: |
    SEAM (Self-documenting Endpoint Access Mediator) is a unified gateway that proxies
    authenticated service requests while injecting required secrets. This spec describes
    the Phase 1a control-plane endpoints.

    **Phase 1a scope:** Gateway scaffold with hand-written whole spec (no fragment merge yet).
    Control-plane endpoints are reserved and short-circuit route-table lookup.
  contact:
    name: SEAM Project
    url: https://github.com/ardenone/seam
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# servers will be populated at runtime with the caller-facing base URL
servers: []
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error response format for all API errors
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: >-
            Machine-readable error code identifier. This enum covers every code
            SEAM can return in a top-level ErrorResponse. Fan-out requests are
            the one exception: a 207 Multi-Status envelope reports per-instance
            outcomes through the hyphenated `status` vocabulary
            ("ok", "error", "timeout", "breaker-refused", "truncated",
            "scope-withheld") carried by each entry in `instances`, not through
            this field.
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - invalid_version_parameter
            - missing_required_parameter
            - invalid_payload
            - route_not_found
            - validation_failed
            - quota_exceeded
            - rate_limit_exceeded
            - loop_guard_exceeded
            - internal_server_error
            - bad_gateway
            - service_unavailable
            - gateway_timeout
            - upstream_failed
            - no_upstream_configured
            - proxy_creation_failed
            - capture_failed
            - spec_load_failed
            - config_error
            - credential_refresh_not_retried
            - secret_store_unavailable
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          description: Caller-safe error context and metadata; internal causes are never exposed
          additionalProperties: true
        validation_errors:
          type: array
          description: Field-level diagnostics, present only when error is validation_failed
          items:
            $ref: '#/components/schemas/ValidationError'
        docs_url:
          type: string
          description: Link to relevant documentation
          format: uri-reference
        request_id:
          type: string
          description: Unique request identifier for tracing
      example:
        error: route_not_found
        message: No route found for GET /unknown/path
        details:
          method: GET
          path: /unknown/path
        docs_url: /docs
        request_id: req_123456

    # Validation errors use the same envelope and add field-level diagnostics.
...
    ValidationError:
      type: object
      description: Field-level validation error
      required:
        - field
        - expected_shape
        - reason
      properties:
        field:
          type: string
          description: Field or parameter name that failed validation
        expected_shape:
          type: string
          description: Expected value or format for the field
        actual:
          type: string
          description: Actual value received (optional)
        reason:
          type: string
          description: Why validation failed
        line:
          type: integer
          minimum: 0
          description: OpenAPI source line when supplied by the validator
        column:
          type: integer
          minimum: 0
          description: OpenAPI source column when supplied by the validator

  # Reusable response templates
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters or payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: bad_request
            message: Invalid request parameters
            details:
              parameter: version
...
    NotFound:
      description: Not Found - Resource or route does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: route_not_found
            message: No route found for GET /unknown/path
            docs_url: /docs
    MethodNotAllowed:
      description: Method Not Allowed - HTTP method not supported for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: method_not_allowed
            message: Method not allowed
...
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: internal_server_error
            message: An unexpected error occurred
...
paths:
...
  /docs:
    get:
      summary: API documentation
      description: Returns rendered API documentation (HTML or JSON)
      operationId: getDocs
      tags:
        - Documentation
      responses:
        '200':
          description: API documentation
          content:
            application/json:
              schema:
                type: object
            text/html:
              schema:
                type: string
          headers:
            X-SEAM-Spec-Version:
              description: Stable hash of the served spec
              schema:
                type: string
            X-SEAM-API-Version:
              description: API version
              schema:
                type: string
                default: _unversioned
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /docs/route:
    get:
      summary: Get documentation for one route
      description: |
        Returns the served OpenAPI slice for the path template in `path`.
        Omit `method` to receive every operation declared on that path. The
        route selector is a query parameter so encoded slashes in templates
        remain normalization-safe through the caller-facing stack.
      operationId: getRouteDocs
      tags:
        - Documentation
      parameters:
        - name: path
          in: query
          required: true
          description: OpenAPI path template, such as /widgets/{id}
          schema:
            type: string
          example: /openapi.json
        - name: method
          in: query
          required: false
          description: HTTP method to select; omit for every method on the path
          schema:
            type: string
            enum: [GET, PUT, POST, DELETE, PATCH, OPTIONS, HEAD, TRACE]
          example: GET
        - name: version
          in: query
          required: false
          description: Served API contract version (the only Phase 1a contract)
          schema:
            type: string
            enum: [_unversioned]
            default: _unversioned
      responses:
        '200':
          description: Route parameters, schemas, annotations, and worked example
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...